feat(Circle): Add motion pause functionality
A new property `motionPaused` is added to the `Circle` class. This
property is used to pause the motion of the circles. The `update`
method of the `Circle` class is modified to check the `motionPaused`
property before updating the circle's position. The `createCircles`
function is also updated to accept a new argument `isMotionPaused`
which is passed to the `Circle` constructor.